feat : utils 앱 추가 및 코드 변경
- 여러 앱에서 사용할 코드들의 재활용을 위해 utils 앱으로 묶음
- 테스트가 끝난 뒤 merge 해야할듯.
