+++ quotes/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path, include
 from rest_framework import routers
-from .views import QuoteViewSet, QuoteCommentViewSet, search_address
+from .views import QuoteViewSet, QuoteCommentViewSet
 
 router = routers.DefaultRouter()
 router.register('', QuoteViewSet, basename='quote')
@@ -8,5 +8,4 @@ router.register(r'(?P<quote_pk>\d+)/comments', QuoteCommentViewSet, basename='qu
 
 urlpatterns = [
     path('', include(router.urls)),
-    path('search/address/', search_address, name='search_address'),
 ]
