+++ utils/models.py
@@ -1,0 +1,46 @@
+from django.db import models
+from django.conf import settings
+from django.core.validators import MinValueValidator, MaxValueValidator
+from accounts.models import Company
+
+# Create your models here.
+
+STATUS_CHOICES = (
+        ('MATCHING', '매칭전'),
+        ('MATCHED', '매칭완료'),
+        ('DEPOSIT', '입금완료'),
+        ('PREPARING', '준비중'),
+        ('COMPLETED', '완료'),
+    )
+
+class Article(models.Model):
+    customer = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
+    company = models.ForeignKey(settings.AUTH_USER_MODEL, related_name='quotes_as_company', null=True, blank=True, on_delete=models.CASCADE)
+    title = models.CharField(max_length=30)
+    content = models.TextField()
+    created_at = models.DateTimeField(auto_now_add=True)
+    date = models.DateTimeField()
+    status = models.CharField(max_length=20, choices=STATUS_CHOICES, default='MATCHING')
+
+    class Meta:
+        abstract = True
+
+class Comment(models.Model):
+    author = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
+    amount = models.IntegerField()
+    created_at = models.DateTimeField(auto_now_add=True)
+    is_read = models.BooleanField(default=False)
+
+    class Meta:
+        abstract = True
+        ordering = ['amount']
+
+class Review(models.Model):
+    company = models.ForeignKey(Company, on_delete=models.CASCADE, related_name='reviews')
+    author = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
+    rating = models.IntegerField(validators=[MinValueValidator(1), MaxValueValidator(5)])
+    comment = models.TextField(max_length=500, blank=True)
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    class Meta:
+        abstract = True
